scripts/datagen/requirements.txt
confluent-kafka
python-dotenv
tinkoff-investments
pandas